+++ src/commands/pickFuncProcess.ts
@@ -5,15 +5,13 @@
 
 import { sendRequestWithTimeout, type AzExtRequestPrepareOptions } from '@microsoft/vscode-azext-azureutils';
 import { callWithTelemetryAndErrorHandling, parseError, UserCancelledError, type IActionContext } from '@microsoft/vscode-azext-utils';
-import * as fs from 'fs';
-import * as os from 'os';
-import * as path from 'path';
 import psTree, { type PS } from 'ps-tree';
 import * as vscode from 'vscode';
-import { dotnetIsolatedDebugFlag, enableJsonOutputFlag, hostStartTaskName, jsonOutputFileFlag } from '../constants';
+import { hostStartTaskName, jsonOutputFileFlag } from '../constants';
 import { preDebugValidate, type IPreDebugValidateResult } from '../debug/validatePreDebug';
 import { ext } from '../extensionVariables';
 import { buildPathToWorkspaceFolderMap, getFuncPortFromTaskOrProject, isFuncHostTask, runningFuncTaskMap, stopFuncTaskIfRunning, type IRunningFuncTask } from '../funcCoreTools/funcHostTask';
+import { generateJsonOutputFilePath, getWorkerPidFromJsonOutput, injectJsonOutputFileArgIfNeeded, isDotnetIsolatedDebugTask, shouldInjectJsonOutputFile } from '../funcCoreTools/jsonOutputFile';
 import { localize } from '../localize';
 import { delay } from '../utils/delay';
 import { requestUtils } from '../utils/requestUtils';
@@ -195,7 +193,7 @@ async function startFuncTask(context: IActionContext, workspaceFolder: vscode.Wo
                     // Prefer the file written by func core tools via --json-output-file; if that flag
                     // isn't present (e.g., the task was already running before we could inject it),
                     // fall back to the worker PID parsed from the terminal stream by funcHostTask.
-                    const newPid = (taskInfo.workerPidFile ? await getWorkerPidFromJsonOutput(taskInfo) : undefined)
+                    const newPid = await getWorkerPidFromJsonOutput(taskInfo.workerPidFile)
                         ?? taskInfo.workerProcessId;
                     if (newPid) {
                         taskInfo.processId = newPid;
@@ -236,109 +234,6 @@ async function startFuncTask(context: IActionContext, workspaceFolder: vscode.Wo
     } finally {
         errorListener.dispose();
     }
-}
-
-async function getWorkerPidFromJsonOutput(taskInfo: IRunningFuncTask): Promise<number | undefined> {
-    // func core tools writes the worker PID JSON to the path passed via --json-output-file before
-    // the worker waits for the debugger to attach. Read that file and parse out workerProcessId.
-    if (!taskInfo.workerPidFile) {
-        return;
-    }
-
-    try {
-        if (!fs.existsSync(taskInfo.workerPidFile)) {
-            return;
-        }
-        const content = fs.readFileSync(taskInfo.workerPidFile, 'utf8');
-        const obj = JSON.parse(content) as Record<string, unknown>;
-        if (typeof obj['workerProcessId'] === 'number') {
-            return obj['workerProcessId'];
-        }
-    } catch {
-        // file not yet written or invalid JSON - keep waiting
-    }
-    return;
-}
-
-/**
- * Returns true if `commandOrArgs` contains the flags that indicate the func host is being launched
- * for dotnet-isolated debugging and is configured to emit JSON output. When both are present, we
- * inject `--json-output-file <path>` so that func core tools writes the worker PID directly to a
- * file, allowing pickFuncProcess to attach the debugger before any user code runs.
- */
-function shouldInjectJsonOutputFile(commandLine: string): boolean {
-    return commandLine.includes(dotnetIsolatedDebugFlag)
-        && commandLine.includes(enableJsonOutputFlag)
-        && !commandLine.includes(jsonOutputFileFlag);
-}
-
-/**
- * Returns true if the func task is configured for dotnet-isolated debugging with JSON output. When
- * true, the picker waits for a worker PID (from --json-output-file or parsed from the terminal
- * stream) instead of polling the host status endpoint.
- */
-function isDotnetIsolatedDebugTask(funcTask: vscode.Task): boolean {
-    if (!(funcTask.execution instanceof vscode.ShellExecution)) {
-        return false;
-    }
-    const flatArgs = getFlatShellArgs(funcTask.execution);
-    return flatArgs.includes(dotnetIsolatedDebugFlag) && flatArgs.includes(enableJsonOutputFlag);
-}
-
-function generateJsonOutputFilePath(): string {
-    return path.join(os.tmpdir(), `azfunc-worker-pid-${process.pid}-${Date.now()}.json`);
-}
-
-/**
- * If the user's func task is configured for dotnet-isolated debugging with JSON output but does not
- * already specify a `--json-output-file`, returns a wrapper task with that flag injected so that
- * func core tools writes the worker PID directly to a file we control.
- */
-function injectJsonOutputFileArgIfNeeded(funcTask: vscode.Task): vscode.Task {
-    const exec = funcTask.execution;
-    if (!(exec instanceof vscode.ShellExecution)) {
-        return funcTask;
-    }
-
-    const flatArgs = getFlatShellArgs(exec);
-    const hasDebugFlag = flatArgs.includes(dotnetIsolatedDebugFlag);
-    const hasEnableJsonOutput = flatArgs.includes(enableJsonOutputFlag);
-    const alreadyHasOutputFile = flatArgs.includes(jsonOutputFileFlag) || flatArgs.some(a => a.startsWith(`${jsonOutputFileFlag}=`));
-    if (!hasDebugFlag || !hasEnableJsonOutput || alreadyHasOutputFile) {
-        return funcTask;
-    }
-
-    const jsonOutputFile = generateJsonOutputFilePath();
-    let newExec: vscode.ShellExecution;
-    if (exec.commandLine !== undefined) {
-        newExec = new vscode.ShellExecution(`${exec.commandLine} ${jsonOutputFileFlag} "${jsonOutputFile}"`, exec.options);
-    } else {
-        // When constructed with command + args, both are defined; defensively coalesce to satisfy the API types.
-        newExec = new vscode.ShellExecution(exec.command ?? 'func', [...(exec.args ?? []), jsonOutputFileFlag, jsonOutputFile], exec.options);
-    }
-
-    const wrapped = new vscode.Task(
-        funcTask.definition,
-        funcTask.scope ?? vscode.TaskScope.Workspace,
-        funcTask.name,
-        funcTask.source,
-        newExec,
-        funcTask.problemMatchers,
-    );
-    wrapped.isBackground = funcTask.isBackground;
-    wrapped.presentationOptions = funcTask.presentationOptions;
-    wrapped.group = funcTask.group;
-    wrapped.runOptions = funcTask.runOptions;
-    wrapped.detail = funcTask.detail;
-    return wrapped;
-}
-
-function getFlatShellArgs(exec: vscode.ShellExecution): string[] {
-    if (exec.commandLine !== undefined) {
-        // Best-effort split for detection; quoting/escaping is preserved in the original commandLine when we re-emit.
-        return exec.commandLine.split(/\s+/).filter(Boolean);
-    }
-    return (exec.args ?? []).map(a => (typeof a === 'string' ? a : a.value));
 }
 
 type OSAgnosticProcess = { command: string | undefined; pid: number | string };
